Fix CDC issue in async FIFO: ensure that Gray code pointers are registered before resynchronisation.
Feeding a binary counter into a Gray encoder produces static hazards on
the encoder outputs, since each XOR gate in the encoder has multiple inputs,
and potentially ALL of these may transition simultaneously, e.g.
when the binary counter rolls over.

Registering the Gray output ensures that only 1 bit of the domain-crossing
path changes at a time. This has been done in a way that does not affect the
timing or operation of the FIFO.