Record: force name to be given by name for `__init__()`
This is to make it easier to support the `like()` class method in subclasses
of Record. If in the subclass the second argument to `__init__()` is not
name one currently has to take into account that the second argument can
actually be the name if the first argument is of type `Layout`.

For example if you have:

    class JTAG(Record):
        def __init__(master=False, trst=False, *, name=None, src_loc_at=0):

If one uses the .like() function on a JTAG object it will pass name in trst
and not in name without this patch.